LayoutTests/batterystatus/basic-all-types-of-events-expected.txt

Tests the basic operation of all BatteryStatus events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


chargingchange event is raised
PASS battery.charging is charging
PASS battery.chargingTime is chargingTime
PASS battery.dischargingTime is dischargingTime
PASS battery.level is level
chargingtimechange event is raised
PASS battery.charging is charging
PASS battery.chargingTime is chargingTime
PASS battery.dischargingTime is dischargingTime
PASS battery.level is level
dischargingtimechange event is raised
PASS battery.charging is charging
PASS battery.chargingTime is chargingTime
PASS battery.dischargingTime is dischargingTime
PASS battery.level is level
levelchange event is raised
PASS battery.charging is charging
PASS battery.chargingTime is chargingTime
PASS battery.dischargingTime is dischargingTime
PASS battery.level is level
PASS successfullyParsed is true

TEST COMPLETE
